+++ surprise/prediction_algorithms/baseline_only.py
@@ -1,5 +1,4 @@
 """
-Algorithm predicting only the baseline.
 """
 
 from __future__ import (absolute_import, division, print_function,